feat: create esm.mjs
